refactor(*): move `GameChatClient` to `clients`
Ref: #574
