+++ src-frontend-app-web/src/clients/index.ts
@@ -1,6 +1,7 @@
 export * from "./character-creation.client";
 export * from "./configuration.client";
 export * from "./game-action.client";
+export * from "./game-chat.client";
 export * from "./hall-of-fame.client";
 export * from "./localization.client";
 export * from "./login.client";
